requirements.txt
torch
transformers
bert-extractive-summarizer
spacy
openai-whisper
python-dotenv
streamlit
accelerate
sentencepiece
numpy